Silence compiler warnings.
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@22846 154be670-7969-47af-9ab0-c6fb378d2295

+++ ext/CrashServer/CrashHandler/DumpUploaderServiceLib/DumpUploaderWebService.cpp
@@ -70,7 +70,7 @@ std::wstring DumpUploaderWebService::GetErrorText()
 	std::ostringstream o;
 	soap_stream_fault(o); 
 	std::vector<wchar_t> buf(o.str().size() + 1);
-	buf.resize(MultiByteToWideChar(CP_UTF8, 0, (LPCSTR) o.str().c_str(), o.str().size() + 1, &buf.front(), buf.size()));
+	buf.resize(MultiByteToWideChar(CP_UTF8, 0, (LPCSTR) o.str().c_str(), (int)o.str().size() + 1, &buf.front(), (int)buf.size()));
 
 	return std::wstring(buf.begin(), buf.end());
 }
